[+] Split schemas - Update pre-commit add template for apps
