+++ .pre-commit-config.yaml
@@ -21,12 +21,14 @@ repos:
     rev: 6.1.0  # Latest version
     hooks:
       - id: flake8
+        args: ['--max-line-length=120']
         additional_dependencies: [flake8-bugbear]
 
   - repo: https://github.com/PyCQA/isort
     rev: 5.12.0  # Latest version
     hooks:
       - id: isort
+        args : ['--profile=black']
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
   # Ruff version.
@@ -48,8 +50,9 @@ repos:
     rev: v1.11.2
     hooks:
       - id: mypy
-        additional_dependencies: [sqlalchemy-stubs, types-ujson]
+        files: src/.*
         args: ['--ignore-missing-imports']
+
   - repo: https://github.com/asottile/pyupgrade
     rev: v3.17.0
     hooks:
