+++ src/cutsom_types.py
@@ -1,0 +1,6 @@
+from typing import NewType
+
+HashedPassword = NewType("HashedPassword", str)
+PhoneNumber = NewType("PhoneNumber", int)
+NationalID = NewType("NationalID", int)
+StudnentID = NewType("StudnentID", int)
