+++ src/instructor/exceptions.py
@@ -1,0 +1,2 @@
+class InstructorTimeIsFull(BaseException):
+    pass
