+++ src/models.py
@@ -1,0 +1,3 @@
+from sqlalchemy.orm import declarative_base
+
+BaseModel = declarative_base()
